requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
py-cord
python-dotenv
aiosqlite
ezcord
requests
deepl
openai
fastapi>=0.95.2
uvicorn>=0.22
pydantic>=2.0